+++ src/TestSuite/ConsoleIntegrationTestCase.php
@@ -9,20 +9,27 @@
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/cakephp-database-backup
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @since       2.0.0
+ * @since       2.2.0
  */
 namespace DatabaseBackup\TestSuite;
 
+use Cake\Console\Shell;
 use Cake\Core\Configure;
-use Cake\TestSuite\TestCase as CakeTestCase;
+use Cake\TestSuite\ConsoleIntegrationTestCase as CakeConsoleIntegrationTestCase;
 use DatabaseBackup\TestSuite\TestCaseTrait;
+use DatabaseBackup\Utility\BackupExport;
 
 /**
- * TestCase class
+ * ConsoleIntegrationTestCase class
  */
-abstract class TestCase extends CakeTestCase
+class ConsoleIntegrationTestCase extends CakeConsoleIntegrationTestCase
 {
     use TestCaseTrait;
+
+    /**
+     * @var \DatabaseBackup\Utility\BackupExport
+     */
+    protected $BackupExport;
 
     /**
      * Setup the test case, backup the static object values so they can be
@@ -33,6 +40,8 @@ abstract class TestCase extends CakeTestCase
     public function setUp()
     {
         parent::setUp();
+
+        $this->BackupExport = new BackupExport;
 
         $this->deleteAllBackups();
     }
@@ -49,5 +58,27 @@ abstract class TestCase extends CakeTestCase
         $this->deleteAllBackups();
 
         Configure::write(DATABASE_BACKUP . '.connection', 'test');
+    }
+
+    /**
+     * Asserts shell exited with the error code
+     * @param string $message Failure message to be appended to the generated
+     *  message
+     * @return void
+     */
+    public function assertExitWithError($message = '')
+    {
+        $this->assertExitCode(Shell::CODE_ERROR, $message);
+    }
+
+    /**
+     * Asserts shell exited with the success code
+     * @param string $message Failure message to be appended to the generated
+     *  message
+     * @return void
+     */
+    public function assertExitWithSuccess($message = '')
+    {
+        $this->assertExitCode(Shell::CODE_SUCCESS, $message);
     }
 }