+++ src/TestSuite/DriverTestCase.php
@@ -256,7 +256,7 @@ abstract class DriverTestCase extends TestCase
     public function testExportAndImport()
     {
         foreach (VALID_EXTENSIONS as $extension) {
-            $this->loadAllFixtures();
+            $this->loadFixtures();
 
             $backup = $this->getAbsolutePath(sprintf('example.%s', $extension));